+++ CHANGELOG.md
@@ -211,6 +211,10 @@ Indexes iterators names has been shortened to `Iter`, `Keys` and `Values`. (#162
 
 - Slash (`/`) is now allowed to be a part of artifact/instance name. (#1681)
 
+#### exonum-cli
+
+- Added maintenance command `restart-migration` to restart migration script. (#1728)
+
 #### exonum-merkledb
 
 - MerkleDB now performs automated state aggregation allowing to construct proofs